move mach.rb to os/mac/mach.rb
