+++ Library/Homebrew/os/mac/pathname.rb
@@ -1,0 +1,5 @@
+require "os/mac/mach"
+
+class Pathname
+  include MachO
+end
